First fully functional version.
Once the program starts, it will check http://block.hardchats.com/~asterisk/voice_recordings/ periodically for new audio files. The frequency of this checking can be controlled with the UPDATE_DOWNLOAD_INTERVAL global variable.

Each message will be played a number of times which is inversely proportional to the size of the message queue. There is also a delay between two consecutive plays, which is proportional to the size of the queue. Other more sophisticated approaches could be implemented to determine these two variables (number of playbacks, and pause time between them). The function that contains this logic is Message.set_playback()

Status text could be shown and updated live on the main pyglet output window, but I leave this for later. Right now there are a couple of keys that can be used to control some basic features of the program (space to print info, up/down to adjust playback volume, esc to quit the program). Try to get it running and let me know if there are problems.